+++ src/tools/context.ts
@@ -1,3 +1,0 @@
-export enum Context {
-  CandyData = "candy-data",
-}