+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [v1.5.0](https://github.com/Candy-Doc/candy-board/compare/v1.4.0...v1.5.0) (2023-06-05)
+
+## 💄 UI and style files added or updated
+- [`9223be8`](https://github.com/Candy-Doc/candy-board/commit/9223be8)  style: implement new sidebar from candy-doc/ui update (#37) (Issues: [`#37`](https://github.com/Candy-Doc/candy-board/issues/37))
+
+# [v1.4.0](https://github.com/Candy-Doc/candy-board/compare/v1.3.0...v1.4.0) (2023-06-05)
+
+## 💄 UI and style files added or updated
+- [`4aabe6b`](https://github.com/Candy-Doc/candy-board/commit/4aabe6b)  UI: implement new topbar from candy-doc/ui update (#35) (Issues: [`#35`](https://github.com/Candy-Doc/candy-board/issues/35))
+
+# [v1.3.0](https://github.com/Candy-Doc/candy-board/compare/v1.2.0...v1.3.0) (2023-05-31)
+
+## ✨ New minor Features
+- [`33dbd42`](https://github.com/Candy-Doc/candy-board/commit/33dbd42)  feat: new design (#34) (Issues: [`#34`](https://github.com/Candy-Doc/candy-board/issues/34))
+
 # [v1.2.0](https://github.com/Candy-Doc/candy-board/compare/v1.1.3...v1.2.0) (2023-05-17)
 
 ## ✨ New minor Features